requirements.txt
httpx==0.27.0
pandas==2.2.2
tqdm==4.66.4
python-dotenv==1.0.1
openpyxl==3.1.5
beautifulsoup4==4.12.3
lxml==5.2.2
ollama==0.4.7
streamlit==1.36.0
plotly==5.22.0
